whole 5 videos are now working!

+++ robot/models.py
@@ -13,7 +13,7 @@ class TelegramUser(models.Model):
     chat_id = models.CharField(
         max_length=20
     )
-    balance = models.IntegerField(
+    balance = models.FloatField(
         default=0
     )
